+++ pandas/core/indexes/timedeltas.py
@@ -1,7 +1,5 @@
 """ implement the TimedeltaIndex """
 
-from datetime import timedelta
-
 import numpy as np
 from pandas.core.dtypes.common import (
     _TD_DTYPE,
@@ -17,7 +15,8 @@ from pandas.core.dtypes.common import (
 from pandas.core.dtypes.missing import isna
 from pandas.core.dtypes.generic import ABCSeries
 
-from pandas.core.arrays.timedelta import TimedeltaArrayMixin
+from pandas.core.arrays.timedelta import (
+    TimedeltaArrayMixin, _is_convertible_to_td)
 from pandas.core.indexes.base import Index
 from pandas.core.indexes.numeric import Int64Index
 import pandas.compat as compat
@@ -33,7 +32,7 @@ from pandas.core.indexes.datetimelike import (
     TimelikeOps, DatetimeIndexOpsMixin)
 from pandas.core.tools.timedeltas import (
     to_timedelta, _coerce_scalar_to_timedelta_type)
-from pandas.tseries.offsets import Tick, DateOffset
+from pandas.tseries.offsets import DateOffset
 from pandas._libs import (lib, index as libindex,
                           join as libjoin, Timedelta, NaT, iNaT)
 
@@ -268,46 +267,10 @@ class TimedeltaIndex(TimedeltaArrayMixin, DatetimeIndexOpsMixin,
 
     @classmethod
     def _generate(cls, start, end, periods, name, freq, closed=None):
-        if com._count_not_none(start, end, periods, freq) != 3:
-            raise ValueError('Of the four parameters: start, end, periods, '
-                             'and freq, exactly three must be specified')
-
-        if start is not None:
-            start = Timedelta(start)
-
-        if end is not None:
-            end = Timedelta(end)
-
-        left_closed = False
-        right_closed = False
-
-        if start is None and end is None:
-            if closed is not None:
-                raise ValueError("Closed has to be None if not both of start"
-                                 "and end are defined")
-
-        if closed is None:
-            left_closed = True
-            right_closed = True
-        elif closed == "left":
-            left_closed = True
-        elif closed == "right":
-            right_closed = True
-        else:
-            raise ValueError("Closed has to be either 'left', 'right' or None")
-
-        if freq is not None:
-            index = _generate_regular_range(start, end, periods, freq)
-            index = cls._simple_new(index, name=name, freq=freq)
-        else:
-            index = to_timedelta(np.linspace(start.value, end.value, periods))
-
-        if not left_closed:
-            index = index[1:]
-        if not right_closed:
-            index = index[:-1]
-
-        return index
+        # TimedeltaArray gets `name` via **kwargs, so we need to explicitly
+        # override it if name is passed as a positional argument
+        return super(TimedeltaIndex, cls)._generate(start, end, periods, freq,
+                                                    name=name, closed=closed)
 
     @classmethod
     def _simple_new(cls, values, name=None, freq=None, **kwargs):
@@ -383,90 +346,8 @@ class TimedeltaIndex(TimedeltaArrayMixin, DatetimeIndexOpsMixin,
     microseconds = _wrap_field_accessor("microseconds")
     nanoseconds = _wrap_field_accessor("nanoseconds")
 
-    @property
-    def components(self):
-        """
-        Return a dataframe of the components (days, hours, minutes,
-        seconds, milliseconds, microseconds, nanoseconds) of the Timedeltas.
-
-        Returns
-        -------
-        a DataFrame
-        """
-        from pandas import DataFrame
-
-        columns = ['days', 'hours', 'minutes', 'seconds',
-                   'milliseconds', 'microseconds', 'nanoseconds']
-        hasnans = self.hasnans
-        if hasnans:
-            def f(x):
-                if isna(x):
-                    return [np.nan] * len(columns)
-                return x.components
-        else:
-            def f(x):
-                return x.components
-
-        result = DataFrame([f(x) for x in self])
-        result.columns = columns
-        if not hasnans:
-            result = result.astype('int64')
-        return result
-
+    @Appender(TimedeltaArrayMixin.total_seconds.__doc__)
     def total_seconds(self):
-        """
-        Return total duration of each element expressed in seconds.
-
-        This method is available directly on TimedeltaIndex and on Series
-        containing timedelta values under the ``.dt`` namespace.
-
-        Returns
-        -------
-        seconds : Float64Index or Series
-            When the calling object is a TimedeltaIndex, the return type is a
-            Float64Index. When the calling object is a Series, the return type
-            is Series of type `float64` whose index is the same as the
-            original.
-
-        See Also
-        --------
-        datetime.timedelta.total_seconds : Standard library version
-            of this method.
-        TimedeltaIndex.components : Return a DataFrame with components of
-            each Timedelta.
-
-        Examples
-        --------
-        **Series**
-
-        >>> s = pd.Series(pd.to_timedelta(np.arange(5), unit='d'))
-        >>> s
-        0   0 days
-        1   1 days
-        2   2 days
-        3   3 days
-        4   4 days
-        dtype: timedelta64[ns]
-
-        >>> s.dt.total_seconds()
-        0         0.0
-        1     86400.0
-        2    172800.0
-        3    259200.0
-        4    345600.0
-        dtype: float64
-
-        **TimedeltaIndex**
-
-        >>> idx = pd.to_timedelta(np.arange(5), unit='d')
-        >>> idx
-        TimedeltaIndex(['0 days', '1 days', '2 days', '3 days', '4 days'],
-                       dtype='timedelta64[ns]', freq=None)
-
-        >>> idx.total_seconds()
-        Float64Index([0.0, 86400.0, 172800.0, 259200.00000000003, 345600.0],
-                     dtype='float64')
-        """
         result = TimedeltaArrayMixin.total_seconds(self)
         return Index(result, name=self.name)
 
@@ -915,11 +796,6 @@ def _is_convertible_to_index(other):
     return False
 
 
-def _is_convertible_to_td(key):
-    return isinstance(key, (Tick, timedelta,
-                            np.timedelta64, compat.string_types))
-
-
 def _to_m8(key):
     """
     Timedelta-like => dt64
@@ -930,28 +806,6 @@ def _to_m8(key):
 
     # return an type that can be compared
     return np.int64(key.value).view(_TD_DTYPE)
-
-
-def _generate_regular_range(start, end, periods, offset):
-    stride = offset.nanos
-    if periods is None:
-        b = Timedelta(start).value
-        e = Timedelta(end).value
-        e += stride - e % stride
-    elif start is not None:
-        b = Timedelta(start).value
-        e = b + periods * stride
-    elif end is not None:
-        e = Timedelta(end).value + stride
-        b = e - periods * stride
-    else:
-        raise ValueError("at least 'start' or 'end' should be specified "
-                         "if a 'period' is given.")
-
-    data = np.arange(b, e, stride, dtype=np.int64)
-    data = TimedeltaIndex._simple_new(data, None)
-
-    return data
 
 
 def timedelta_range(start=None, end=None, periods=None, freq=None,
